+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Compatibility:
 * Support an `Array` of `Thread::Backtrace::Location` in `Exception#set_backtrace` and `{Kernel,Thread,Fiber}#raise` (#3883, @eregon).
 * Try `#to_ary` coercion in `Array#assoc` and `Array#rassoc` (#4137, @herwinw).
 * Improve compatibility of `rb_scan_args()` with Ruby 3+ changes (@eregon).
+* Implement `Time#{xmlschema,iso8601}` in core like Ruby 3.4 (#3883, @eregon).
 
 Performance: